scripts/phemex/requirements.txt
ccxt
pandas
pandas_ta
python-dotenv
schedule

